fix: 修复 ExtensionColorDao 与 FavoriteDao 数据库 API 调用并消除类型转换警告
- 使用 DatabaseManager::instance().getGlobalDb() 与 DatabaseManager::instance().getGlobalMutex() 替换不存在的静态方法
- 使用原生 sqlite3_wal_checkpoint_v2 替换不存在的 flushWalCheckpoint
- 在 ColorPaletteEngine.cpp 中显式转换 static_cast<uint>(qHash(e)) 消除 C4267 警告

+++ src/util/ColorPaletteEngine.cpp
@@ -96,7 +96,7 @@ QPair<QColor, QColor> ColorPaletteEngine::getExtensionBadgeColors(const QString&
     }
 
     // 3. 动态生成新配色并物理落盘写入 global.db
-    uint hashVal = qHash(e);
+    uint hashVal = static_cast<uint>(qHash(e));
     int hue = static_cast<int>(hashVal % 360);
     int saturation = 130 + static_cast<int>((hashVal >> 8) % 80);
     int lightness = 80 + static_cast<int>((hashVal >> 16) % 60);
